+++ cascade/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+
+* Add a `c:Cascade.Template.post_generate/1` callback to the `Cascade.Template`
+behaviour.
+
 ## [v0.1.0](https://github.com/sportradar/elixir-workspace/tree/cascade/v0.1.0) (2024-05-13)
 
 Initial release.